Update rm commands to only delete when a file is present. Add command to run bundle install on each container boot. Use the new Berksfile at the root of the cookbooks dir instead of the one just at the cdo-apps dir.

+++ start.sh
@@ -1,12 +1,15 @@
 #!/bin/bash
 
-sudo rm /var/run/unicorn/dashboard.sock
-sudo rm /var/run/unicorn/pegasus.sock
+[ -f "/var/run/unicorn/dashboard.sock" ] && sudo rm /var/run/unicorn/dashboard.sock
+[ -f "/var/run/unicorn/pegasus.sock" ] && sudo rm /var/run/unicorn/pegasus.sock
 
-sudo rm /home/ubuntu/adhoc/dashboard/config/unicorn.rb.pid
-sudo rm /home/ubuntu/adhoc/pegasus/config/unicorn.rb.pid
+[ -f "/home/ubuntu/adhoc/dashboard/config/unicorn.rb.pid" ] && sudo rm /home/ubuntu/adhoc/dashboard/config/unicorn.rb.pid
+[ -f "/home/ubuntu/adhoc/pegasus/config/unicorn.rb.pid" ] && sudo rm /home/ubuntu/adhoc/pegasus/config/unicorn.rb.pid
 
 sudo service mysql start
+
+cd /home/ubuntu/adhoc/
+sudo bundle install
 
 cd /home/ubuntu/adhoc/dashboard
 sudo bundle exec rake db:migrate
@@ -14,9 +17,10 @@ sudo bundle exec rake db:migrate
 cd /home/ubuntu/adhoc/pegasus
 sudo bundle exec rake db:migrate
 
-rm -rf /tmp/chef/cookbooks
+[ -d "/tmp/chef/cookbooks" ] && rm -rf /tmp/chef/cookbooks
+[ -d "/tmp/chef/local-mode-cache" ] && sudo rm -rf /tmp/chef/local-mode-cache
 mkdir /tmp/chef/cookbooks
-cd /home/ubuntu/adhoc/cookbooks/cdo-apps
+cd /home/ubuntu/adhoc/cookbooks
 berks vendor /tmp/chef/cookbooks
 
 sudo chef-client -z -c /tmp/chef/local-adhoc.rb -j /tmp/chef/local-adhoc.json
